KAMANJA-343 Cluster Component Check
1. InstallDriver has checks for scala and java versions on all cluster nodes using the script
found in src/main/resources. The kafka, zk and hbase are checked for their respective status (LOOKING for "ok" for now)
2. The script and the CheckComponent jar paths are passed in on the command line
3. The lower level script has been modified to accept the physical directory (from which we get the parent path where the
old and new installations will be placed and the generated names for old and new installations.
4. The names are based upon the logical (symbolic link) name in form <Value from ROOT_DIR>_<Old|NewKamanjaVer>_YYYYmmdd_HHmmss
5. The client ips connection strings are dug out of the cluster config.
6. The glue code that Pokuri developed that invokes the GetComponentsVersion.sh script was integrated.
7. The cmd for invoking the lower level install was fixed (commas removed).
8. The unlink and relink code in the lower level installer now uses the physical installs parent directory and the new names instead
of the date suffix.

To do:
1) This builds but I did not test end to end yet. Except has noted in the next few items, the features are there afaict.
2) We need to make sure that the paths of the old and new installations are prominently displayed in the top level log.
3) We need to split the emissions from the lower level cluster installer into really meaningful and detailed where the
detailed echos get appended into a file whose path is also passed to the lower level script.  The stdout will (important
events and failures) go into the high level log.  The detailed echos collected in the file are put in the log4j log.
This work has not been started.
4) The environment check function should be carefully reviewed to see if the returned information is being returned properly.
5) Other checks can and should be developed to eagerly find as many environment related issues as possible.
6) The messages mentioned in the documentation that identify what the prior install is named and what the new physical
install location for new install should be added or synced up with the source code/scripts.
7) I am sure there will be another bunch of things to do as well.

+++ trunk/SampleApplication/ClusterInstall/scala-2.11/KamanjaClusterInstall.sh
@@ -63,7 +63,12 @@ Usage()
     echo "                               --TarballPath <tarball path>"
     echo "                               --NodeConfigPath <engine config path> "
     echo "                               [ --WorkingDir <alt working dir>  ]"
-    echo 
+    echo "                               --ipAddrs the ips in the cluster  "
+    echo "                               --ipIdTargPaths unique ips, node ids, target paths, and roles  "
+    echo "                               --ipPathPairs unique ips and target paths  "
+    echo "                               --priorInstallDirPath name to use for rename of prior install if any  "
+    echo "                               --newInstallDirPath new name of physical install  "
+    echo
     echo "  NOTES: Only tar'd gzip files are supported for the tarballs at the moment."
     echo "         NodeConfigPath must be supplied always"
     echo "         The working directory, by default, is /tmp.  If such a public location is abhorrent, chose a private one.  It"
@@ -80,7 +85,7 @@ Usage()
 
 
 # Check 1: Is this even close to reasonable?
-if [[ "$#" -eq 4  || "$#" -eq 6  || "$#" -eq 8  || "$#" -eq 10 ]]; then
+if [[ "$#" -eq 1  || "$#" -eq 4  || "$#" -eq 6  || "$#" -eq 8  || "$#" -eq 10  || "$#" -eq 12  || "$#" -eq 14  || "$#" -eq 16  || "$#" -eq 18  || "$#" -eq 20 ]]; then
     echo 
 else 
     echo 
@@ -90,9 +95,9 @@ else
 fi
 
 # Check 2: Is this even close to reasonable?
-if [[ "$name1" != "--ClusterId" && "$name1" != "--MetadataAPIConfig" && "$name1" != "--NodeConfigPath"  && "$name1" != "--KafkaInstallPath"   && "$name1" != "--TarballPath"  && "$name1" != "--WorkingDir" ]]; then
+if [[ "$name1" != "--ClusterId" && "$name1" != "--MetadataAPIConfig" && "$name1" != "--NodeConfigPath"  && "$name1" != "--KafkaInstallPath"   && "$name1" != "--TarballPath"  && "$name1" != "--WorkingDir"   && "$name1" != "--ipAddrs"   && "$name1" != "--ipIdTargPaths"   && "$name1" != "--ipPathPairs" && "$name1" != --priorInstallDirPath" &&  n"$name1" != ""--newInstallDirPath" ]]; then
     echo 
-	echo "Problem: Unreasonable number of arguments... as few as 2 and as many as 4 may be supplied."
+	echo "Problem: Unreasonable number of arguments... as few as 2 and as many as 20 may be supplied."
     Usage
 	exit 1
 fi
@@ -106,6 +111,11 @@ nodeCfgGiven=""
 workDir="/tmp"
 installDirName="" 
 clusterId=""
+ipAddrs=""
+ipIdTargPaths=""
+ipPathPairs=""
+priorInstallDirPath=""
+newInstallDirPath=""
 
 while [ "$1" != "" ]; do
     case $1 in
@@ -128,7 +138,25 @@ while [ "$1" != "" ]; do
         --ClusterId )           shift
                                 clusterId=$1
                                 ;;
-        * )                     echo 
+        --ipAddrs )           	shift
+                                ipAddrs=$1
+                                ;;
+        --ipIdTargPaths )       shift
+                                ipIdTargPaths=$1
+                                ;;
+        --ipPathPairs )         shift
+                                ipPathPairs=$1
+                                ;;
+        --priorInstallDirPath ) shift
+                                priorInstallDirPath=$1
+                                ;;
+        --newInstallDirPath )   shift
+                                newInstallDirPath=$1
+                                ;;
+        --help )           		Usage
+        						exit 0
+                                ;;
+        * )                     echo
                                 echo "Problem: Argument $1 is invalid named parameter."
                                 Usage
                                 exit 1
@@ -288,7 +316,18 @@ tarName="$dtPrefix.tgz"
 trunkDir=`pwd` #save the current trunk directory 
 
 installDir=`cat $metadataAPIConfig | grep '[Rr][Oo][Oo][Tt]_[Dd][Ii][Rr]' | sed 's/.*=\(.*\)$/\1/g'`
+installDir=`echo "$installDir" | sed -e 's/\/[\/]*$//'`
 installDirName=`echo $installDir | sed 's/.*\/\(.*\)$/\1/g'`
+installFolder=`echo "$installDir" | sed -e 's/\/[^\/]*$//'`
+
+if [ -d "$installFolder" ]; then 
+	echo "$installFolder is a Folder"
+else
+	echo "$installFolder does not exist"
+	Usage
+	exit 1
+fi
+
 if [ -z "$tarballPath" ]; then
 	# 1 build the installation in the staging directory
 	stagingDir="$workDir/$installDirName"
@@ -314,45 +353,50 @@ else
 fi
 
 # 3) determine which machines and installation directories are to get the build from the metadata and Kamanja config
-# A number of files are produced, all in the working dir.
-ipFile="ip.txt"
-ipPathPairFile="ipPath.txt"
-ipIdCfgTargPathQuartetFileName="ipIdCfgTarg.txt"
+# A number of files are produced, all USED TO BE IN WORKING DIR WHEN THEY WERE created by the NodeInfoExtract.  Now
+# these three files are passed in as full paths from the caller.  NodeInfoExtract is not used.
+ipFile="$ipAddrs"
+ipPathPairFile="$ipIdTargPaths"
+ipIdCfgTargPathQuartetFileName="$ipIdTargPaths"
+#ipFile="ip.txt"
+#ipPathPairFile="ipPath.txt"
+#ipIdCfgTargPathQuartetFileName="ipIdCfgTarg.txt"
 
-echo "...extract node information for the cluster to be installed from the Metadata configuration and optional node information supplied"
-if  [ -n "$nodeCfgGiven" ]; then
-	echo "metadataAPIConfig = $metadataAPIConfig"
-	echo "nodeConfigPath = $nodeConfigPath"
-	echo "workDir = $workDir"
-	echo "ipFile = $ipFile"
-	echo "ipPathPairFile = $ipPathPairFile"
-	echo "ipIdCfgTargPathQuartetFileName = $ipIdCfgTargPathQuartetFileName"
-	echo "installDir = $installDir"
-	echo "clusterId = $clusterId"
-	echo "...Command = NodeInfoExtract-1.0 --MetadataAPIConfig \"$metadataAPIConfig\" --NodeConfigPath \"$nodeConfigPath\"  --workDir \"$workDir\" --ipFileName \"$ipFile\" --ipPathPairFileName \"$ipPathPairFile\" --ipIdCfgTargPathQuartetFileName \"$ipIdCfgTargPathQuartetFileName\" --installDir \"$installDir\" --clusterId \"$clusterId\""
-	NodeInfoExtract-1.0 --MetadataAPIConfig "$metadataAPIConfig" --NodeConfigPath "$nodeConfigPath" --workDir "$workDir" --ipFileName "$ipFile" --ipPathPairFileName "$ipPathPairFile" --ipIdCfgTargPathQuartetFileName "$ipIdCfgTargPathQuartetFileName"  --installDir "$installDir" --clusterId "$clusterId"
-	# Check 15: Bad NodeInfoExtract-1.0 arguments
-	if [ "$?" -ne 0 ]; then
-		echo
-		echo "Problem: Invalid arguments supplied to the NodeInfoExtract-1.0 application... unable to obtain node configuration... exiting."
-		Usage
-		exit 1
-	fi
-else # info is assumed to be present in the supplied metadata store... see trunk/utils/NodeInfoExtract for details 
-	echo "...Command = $nodeInfoExtractDir/NodeInfoExtract-1.0 --MetadataAPIConfig \"$metadataAPIConfig\" --workDir \"$workDir\" --ipFileName \"$ipFile\" --ipPathPairFileName \"$ipPathPairFile\" --ipIdCfgTargPathQuartetFileName \"$ipIdCfgTargPathQuartetFileName\" --installDir \"$installDir\" --clusterId \"$clusterId\""
-		NodeInfoExtract-1.0 --MetadataAPIConfig $metadataAPIConfig --workDir "$workDir" --ipFileName "$ipFile" --ipPathPairFileName "$ipPathPairFile" --ipIdCfgTargPathQuartetFileName "$ipIdCfgTargPathQuartetFileName" --installDir "$installDir" --clusterId "$clusterId"
-	# Check 15: Bad NodeInfoExtract-1.0 arguments
-	if [ "$?" -ne 0 ]; then
-		echo
-		echo "Problem: Invalid arguments supplied to the NodeInfoExtract-1.0 application... unable to obtain node configuration... exiting."
-		Usage
-		exit 1
-	fi
-fi
+#echo "...extract node information for the cluster to be installed from the Metadata configuration and optional node information supplied"
+#if  [ -n "$nodeCfgGiven" ]; then
+#	echo "metadataAPIConfig = $metadataAPIConfig"
+#	echo "nodeConfigPath = $nodeConfigPath"
+#	echo "workDir = $workDir"
+#	echo "ipFile = $ipFile"
+#	echo "ipPathPairFile = $ipPathPairFile"
+#	echo "ipIdCfgTargPathQuartetFileName = $ipIdCfgTargPathQuartetFileName"
+#	echo "installDir = $installDir"
+#	echo "clusterId = $clusterId"
+#	echo "...Command = NodeInfoExtract-1.0 --MetadataAPIConfig \"$metadataAPIConfig\" --NodeConfigPath \"$nodeConfigPath\"  --workDir \"$workDir\" --ipFileName \"$ipFile\" --ipPathPairFileName \"$ipPathPairFile\" --ipIdCfgTargPathQuartetFileName \"$ipIdCfgTargPathQuartetFileName\" --installDir \"$installDir\" --clusterId \"$clusterId\""
+#	NodeInfoExtract-1.0 --MetadataAPIConfig "$metadataAPIConfig" --NodeConfigPath "$nodeConfigPath" --workDir "$workDir" --ipFileName "$ipFile" --ipPathPairFileName "$ipPathPairFile" --ipIdCfgTargPathQuartetFileName "$ipIdCfgTargPathQuartetFileName"  --installDir "$installDir" --clusterId "$clusterId"
+#	# Check 15: Bad NodeInfoExtract-1.0 arguments
+#	if [ "$?" -ne 0 ]; then
+#		echo
+#		echo "Problem: Invalid arguments supplied to the NodeInfoExtract-1.0 application... unable to obtain node configuration... exiting."
+#		Usage
+#		exit 1
+#	fi
+#else # info is assumed to be present in the supplied metadata store... see trunk/utils/NodeInfoExtract for details
+#	echo "...Command = $nodeInfoExtractDir/NodeInfoExtract-1.0 --MetadataAPIConfig \"$metadataAPIConfig\" --workDir \"$workDir\" --ipFileName \"$ipFile\" --ipPathPairFileName \"$ipPathPairFile\" --ipIdCfgTargPathQuartetFileName \"$ipIdCfgTargPathQuartetFileName\" --installDir \"$installDir\" --clusterId \"$clusterId\""
+#		NodeInfoExtract-1.0 --MetadataAPIConfig $metadataAPIConfig --workDir "$workDir" --ipFileName "$ipFile" --ipPathPairFileName "$ipPathPairFile" --ipIdCfgTargPathQuartetFileName "$ipIdCfgTargPathQuartetFileName" --installDir "$installDir" --clusterId "$clusterId"
+#	# Check 15: Bad NodeInfoExtract-1.0 arguments
+#	if [ "$?" -ne 0 ]; then
+#		echo
+#		echo "Problem: Invalid arguments supplied to the NodeInfoExtract-1.0 application... unable to obtain node configuration... exiting."
+#		Usage
+#		exit 1
+#	fi
+#fi
 
 echo "...creating directories to copy the tarball to the machines in this cluster"
 exec 12<&0 # save current stdin
-exec < "$workDir/$ipFile"
+#exec < "$workDir/$ipFile"
+exec < "$ipFile"
 while read LINE; do
     machine=$LINE
     echo "...creating directory $machine:$workDir"
@@ -366,7 +410,8 @@ exec 0<&12 12<&-
 # 4) Push the tarballs to each machine defined in the supplied configuration
 echo "...copy the tarball to the machines in this cluster"
 exec 12<&0 # save current stdin
-exec < "$workDir/$ipFile"
+#exec < "$workDir/$ipFile"
+exec < "$ipFile"
 while read LINE; do
     machine=$LINE
     echo "...copying $tarName to $machine:$workDir/$tarName"
@@ -383,23 +428,29 @@ DATE=`date +%Y%m%d%H%M%S`
 # 5) untar/decompress tarballs there and move them into place
 echo "...for each directory specified on each machine participating in the cluster, untar and decompress the software to $workDir/$installDirName... then move to corresponding target path"
 exec 12<&0 # save current stdin
-exec < "$workDir/$ipPathPairFile"
+#exec < "$workDir/$ipPathPairFile"
+exec < "$ipPathPairFile"
 while read LINE; do
     machine=$LINE
     read LINE
-    targetPath=$LINE
+    targetPath=`echo "$LINE" | sed -e 's/\/[\/]*$//'`
+    targetFolder=`echo "$targetPath" | sed -e 's/\/[^\/]*$//'`
     targetPath_date="$targetPath"_"$DATE"
     echo "Extract the tarball $tarName and copy it to $targetPath iff $workDir/$installDirName != $targetPath"
 	ssh -o StrictHostKeyChecking=no -T $machine  <<-EOF
-	        cd $workDir
-		if [ ! -L $targetPath ]; then
-			mv 	$targetPath "$targetPath"_pre_"$DATE"			
-		else
-			unlink $targetPath
+		if [ -d "$targetFolder" ]; then 
+			cd $workDir
+			if [ ! -L $targetPath ]; then
+				mv 	$targetPath "$priorInstallDirPath
+			else
+				unlink $targetPath
+			fi
+			mkdir -p $newInstallDirPath
+	 		tar xzf $tarName -C $newInstallDirPath --strip-components 1
+			ln -sf  $newInstallDirPath $targetPath
+	 	else
+			echo "$targetFolder is not directory"	
 		fi
-		mkdir -p $targetPath_date
- 		tar xzf $tarName -C $targetPath_date --strip-components 1
-		ln -sf  $targetPath_date $targetPath
 EOF
 done
 exec 0<&12 12<&-
@@ -411,7 +462,8 @@ installDir_repl=$(echo $installDir | sed 's/\//\\\//g')
 # 6) Push the node$nodeId.cfg file to each cluster node's working directory.
 echo "...copy the node$nodeId.cfg & log files to the machines' ($workDir/$installDirName) for this cluster "
 exec 12<&0 # save current stdin
-exec < "$workDir/$ipIdCfgTargPathQuartetFileName"
+#exec < "$workDir/$ipIdCfgTargPathQuartetFileName"
+exec < "$ipIdCfgTargPathQuartetFileName"
 while read LINE; do
     machine=$LINE
     read LINE
@@ -441,7 +493,8 @@ echo
 # 7) Run SetPaths.sh & Copy given Node Config File (Ex: ClusterConfig.json)
 echo "...for each machine set new paths"
 exec 12<&0 # save current stdin
-exec < "$workDir/$ipPathPairFile"
+#exec < "$workDir/$ipPathPairFile"
+exec < "$ipPathPairFile"
 while read LINE; do
 	machine=$LINE
 	read LINE
@@ -460,7 +513,8 @@ echo
 # 8) clean up
 # echo "...clean up "
 # exec 12<&0 # save current stdin
-# exec < "$workDir/$ipPathPairFile"
+### exec < "$workDir/$ipPathPairFile"
+# exec < "$ipPathPairFile"
 # while read LINE; do
 #     machine=$LINE
 #     read LINE
